+++ ml_tools/__init__.py
@@ -1,1 +1,0 @@
-from ._core._schema import FeatureSchema