+++ ml_tools/schema.py
@@ -1,0 +1,5 @@
+from ._core._schema import FeatureSchema
+
+__all__ = [
+    "FeatureSchema"
+]