+++ ml_tools/serde.py
@@ -2,8 +2,6 @@ from ._core._serde import (
     serialize_object_filename,
     serialize_object,
     deserialize_object,
-    serialize_schema,
-    deserialize_schema,
     info
 )
 
@@ -11,6 +9,4 @@ __all__ = [
     "serialize_object_filename",
     "serialize_object",
     "deserialize_object",
-    "serialize_schema",
-    "deserialize_schema"
 ]